requirements.txt
appnope==0.1.0
auditok==0.2.0a0
autopep8==1.5
backcall==0.1.0
certifi==2019.11.28
cffi==1.14.0
cycler==0.10.0
decorator==4.4.1
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.16.0
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.1
pandas==1.0.1
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.6
PySoundFile==0.9.0.post1
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
scipy==1.4.1
six==1.14.0
sounddevice==0.3.14
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.8
yapf==0.29.0
